Fix of skipping re-generating files
Fix, when persons_with_gender_<version> exists, but czechnames_<version>.out not (for example if user killed create_cedar.sh). Next time, when create_cedar.sh was started, (re-)generating of both file was skipped, but second one file was not exist. It was a problem for next steps of execution.